ML/pyproject.toml
[project]
name = "voc_local_analysis"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nunininu", email = "sgcho0907@gmail.com"},
]
dependencies = ["pandas>=2.3.0", "psycopg2-binary>=2.9.10", "gensim>=4.3.3", "sqlalchemy>=2.0.41", "python-dotenv>=1.1.0", "scikit-learn>=1.6.1", "xgboost>=3.0.2", "joblib>=1.5.1", "jinja2>=3.1.6", "networkx>=3.4.2", "python-louvain>=0.16", "matplotlib>=3.10.3", "pyvis>=0.3.2", "seaborn>=0.13.2", "plotly>=6.1.2", "tqdm>=4.67.1", "python-mecab-ko>=1.3.7", "konlpy>=0.6.0", "git+https://github.com/attardi/wikiextractor.git"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
